feat(core): support custom model json map fn injectionToken
The map func is of type DynamicFormControlModelConfigMapFn which passes an instance of the service with the model. this is done in order to support recursive calling fromJSON and other methods for cases like ARRAY. The motivation behind this PR is to truly support custom controls. Currently you can only extend known types and the extension is limited, you cannot really add new options to the json config. this allowes to map model config json to a new model class with custom options and configurations. My use case is a draggable form control which can be enabled from the json.