Update README.md
Hey, my team and I have been using your library for a huge project and it is incredible! so just wanted to say that.

We have encountered an issue providing multiple validators and have been stuck on it for a while. 

We first tried to add **multi** for each provider, and we dived into the dev tools to find out that a function in your code expects a map and not an array (using .has). 

Eventually, we understood that we just need to add it to the key|value pairs of the **DYNAMIC_VALIDATORS**.

I hope other developers will find it easy to add multiple validators!

Thank you for this helpful library! keep on coding